requirements.txt
streamlit>=1.31.0
pymupdf>=1.23.0
qrcode>=7.4.2
Pillow>=10.0.0
supabase>=2.3.0
pandas>=2.0.0
